LayoutTests/svg/dom/path-parser-expected.txt

CONSOLE MESSAGE: line 9: Error: Problem parsing d="M1,2"
CONSOLE MESSAGE: line 9: Error: Problem parsing d="xM1,2"
CONSOLE MESSAGE: line 9: Error: Problem parsing d="M1,2"
CONSOLE MESSAGE: line 9: Error: Problem parsing d="M1,2x"
CONSOLE MESSAGE: line 9: Error: Problem parsing d=""
CONSOLE MESSAGE: line 9: Error: Problem parsing d="x"
CONSOLE MESSAGE: line 9: Error: Problem parsing d="L1,2"
This tests the SVG path parser by parsing and then re-serializing various paths.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS parsePath('M1,2') is 'M1,2'
PASS parsePath('m1,2') is 'M1,2'
PASS parsePath('M100,200 m3,4') is 'M100,200 M103,204'
PASS parsePath('M100,200 L3,4') is 'M100,200 L3,4'
PASS parsePath('M100,200 l3,4') is 'M100,200 L103,204'
PASS parsePath('M100,200 H3') is 'M100,200 L3,200'
PASS parsePath('M100,200 h3') is 'M100,200 L103,200'
PASS parsePath('M100,200 V3') is 'M100,200 L100,3'
PASS parsePath('M100,200 v3') is 'M100,200 L100,203'
PASS parsePath('M100,200 Z') is 'M100,200 Z'
PASS parsePath('M100,200 z') is 'M100,200 Z'
PASS parsePath('M100,200 C3,4,5,6,7,8') is 'M100,200 C3,4,5,6,7,8'
PASS parsePath('M100,200 c3,4,5,6,7,8') is 'M100,200 C103,204,105,206,107,208'
PASS parsePath('M100,200 S3,4,5,6') is 'M100,200 C100,200,3,4,5,6'
PASS parsePath('M100,200 s3,4,5,6') is 'M100,200 C100,200,103,204,105,206'
PASS parsePath('M100,200 Q3,4,5,6') is 'M100,200 C35.3,69.3,3.7,4.7,5,6'
PASS parsePath('M100,200 q3,4,5,6') is 'M100,200 C102,202.7,103.7,204.7,105,206'
PASS parsePath('M100,200 T3,4') is 'M100,200 C100,200,67.7,134.7,3,4'
PASS parsePath('M100,200 t3,4') is 'M100,200 C100,200,101,201.3,103,204'
PASS parsePath('M100,200 A3,4,5,0,0,6,7') is 'M100,200 C141.5,162.8,154.1,89.4,128.2,36.2 C102.2,-17.1,47.5,-30.2,6,7'
PASS parsePath('M100,200 A3,4,5,1,0,6,7') is 'M100,200 C141.5,162.8,154.1,89.4,128.2,36.2 C102.2,-17.1,47.5,-30.2,6,7'
PASS parsePath('M100,200 A3,4,5,0,1,6,7') is 'M100,200 C58.5,237.2,3.8,224.1,-22.2,170.8 C-48.1,117.6,-35.5,44.2,6,7'
PASS parsePath('M100,200 A3,4,5,1,1,6,7') is 'M100,200 C58.5,237.2,3.8,224.1,-22.2,170.8 C-48.1,117.6,-35.5,44.2,6,7'
PASS parsePath('M100,200 a3,4,5,0,0,6,7') is 'M100,200 C98.5,202.3,98.6,205.7,100.2,207.7 C101.9,209.6,104.5,209.3,106,207'
PASS parsePath('M100,200 a3,4,5,0,1,6,7') is 'M100,200 C101.5,197.7,104.1,197.4,105.8,199.3 C107.4,201.3,107.5,204.7,106,207'
PASS parsePath('M100,200 a3,4,5,1,0,6,7') is 'M100,200 C98.5,202.3,98.6,205.7,100.2,207.7 C101.9,209.6,104.5,209.3,106,207'
PASS parsePath('M100,200 a3,4,5,1,1,6,7') is 'M100,200 C101.5,197.7,104.1,197.4,105.8,199.3 C107.4,201.3,107.5,204.7,106,207'
PASS parsePath('M1,2,3,4') is 'M1,2 L3,4'
PASS parsePath('m100,200,3,4') is 'M100,200 L103,204'
PASS parsePath('M 100-200') is 'M100,-200'
PASS parsePath('M 0.6.5') is 'M0.6,0.5'
PASS parsePath(' M1,2') is 'M1,2'
PASS parsePath('  M1,2') is 'M1,2'
PASS parsePath('\tM1,2') is 'M1,2'
PASS parsePath('\nM1,2') is 'M1,2'
PASS parsePath('\rM1,2') is 'M1,2'
PASS parsePath('\vM1,2') is ''
PASS parsePath('xM1,2') is ''
PASS parsePath('M1,2 ') is 'M1,2'
PASS parsePath('M1,2\t') is 'M1,2'
PASS parsePath('M1,2\n') is 'M1,2'
PASS parsePath('M1,2\r') is 'M1,2'
PASS parsePath('M1,2\v') is ''
PASS parsePath('M1,2x') is ''
PASS parsePath('') is ''
PASS parsePath('x') is ''
PASS parsePath('L1,2') is ''
PASS parsePath('M.1 .2 L.3 .4 .5 .6') is 'M0.1,0.2 L0.3,0.4 L0.5,0.6'
PASS successfullyParsed is true

TEST COMPLETE
